+++ SolSpecShell.sh
@@ -16,7 +16,7 @@ read -p "Enter the number of processors [ex. 16]: " num
 
 :: python SolSpec_Call.py $directory $date $wavelength
 
-mpiexec -n $num python Spec_fit_mpi.py $directory $date $wavelength
-:: mpiexec -n $num python Spec_fit_mpi_4d.py $directory $date $wavelength
+:: mpiexec -n $num python Spec_fit_mpi.py $directory $date $wavelength
+mpiexec -n $num python Spec_fit_mpi_4d.py $directory $date $wavelength
 
 :: python SolSpec_Call_Heatmaps.py $directory $date $wavelength