+++ SolSpec_Call.py
@@ -42,8 +42,8 @@ import SolSpec as ss
 #"""
 #directory = 'F:/Users/Brendan/Desktop/SolarProject'
 #directory = '/mnt/data/Gallagher'
-#date = '20130815'
-#wavelength = 193
+#date = '20160426'
+#wavelength = 1600
 #"""
 import sys
 directory = sys.argv[1]
@@ -54,7 +54,7 @@ wavelength = int(sys.argv[3])
 ## create derotated region datacube
 """
 #ss.datacube(directory='F:/Users/Brendan/Desktop/SolarProject/data/20120923/304', date='20120923', wavelength=304, sub_reg_coords=[-528,-132,-100,100], coords_type='arc', bin_frac=1)
-#ss.datacube(directory='%s' % (directory), date='%s' % (date), wavelength= wavelength, sub_reg_coords=[-250,150,-450,-150], coords_type='arc', bin_frac=4)
+#ss.datacube(directory='%s' % (directory), date='%s' % (date), wavelength= wavelength, sub_reg_coords=[-150,250,-30,100], coords_type='arc', bin_frac=1)
 #ss.datacube(directory='%s' % (directory), date='%s' % (date), wavelength= wavelength, sub_reg_coords=[-425,425,-375,375], coords_type='arc', bin_frac=4)
 #ss.datacube(directory='%s' % (directory), date='%s' % (date), wavelength= wavelength, sub_reg_coords=[1950,2950,1000,1600], coords_type='pix', bin_frac=2)
 #ss.datacube(directory='%s' % (directory), date='%s' % (date), wavelength= wavelength, sub_reg_coords=[-528,-132,-100,100], coords_type='arc', bin_frac=1)  #20120923
@@ -67,7 +67,7 @@ wavelength = int(sys.argv[3])
 #TIME = np.load('%s/DATA/Temp/%s/%i/*time.npy')
 #EXPOSURE = np.load('%s/DATA/Temp/%s/%i/*exposure.npy')
 
-#ss.fft_avg(directory='%s' % (directory), date='%s' % (date), wavelength= wavelength, num_seg = 6)  # 3 seg for 20120923, 6 seg for rest
+#ss.fft_avg(directory='%s' % (directory), date='%s' % (date), wavelength= wavelength, num_seg = 12)  # 3 seg for 20120923, 6 seg for rest
 #spectra_array = ss.fft_avg(directory='%s' % (directory), date='%s' % (date), wavelength= '%i' % (wavelength), datacube = DATA, timeseries = TIME, exposure_array = EXPOSURE, num_seg = 6)
 #np.save('F:/Users/Brendan/Desktop/SolarProject/data/20120923/304/20120923_304_-100_100i_-528_-132j_spectra', spectra_array)
 #np.save('C:/Users/Brendan/Desktop/SDO/spectra_20130530_193_2300_2600_2200_3000', spectra)  # now this
@@ -76,7 +76,7 @@ wavelength = int(sys.argv[3])
 """
 ## fft-averaging + 3x3 averaging
 """
-#ss.fft_overlap(directory='%s' % directory, date='%s' % date, wavelength=wavelength, window_length='02:00', overlap_pct=75, pixel_box=False)
+ss.fft_overlap(directory='%s' % directory, date='%s' % date, wavelength=wavelength, window_length='02:00', overlap_pct=75, pixel_box=True)
 
 
 """