+++ Changelog.md
@@ -5,3 +5,17 @@
 - increased parameter array size to 9 to hold r-value
 - changed Spec_fit_mpi.py final time printout format from seconds to hours:minutes:seconds
 - changed a bunch of "for i in range(0,...)" to "for i in range(...)"
+
+4/7 : 
+- added mean, median, +/- standard deviations statistics to histograms
+
+4/8 :
+- added M1 curve fitting + parameter display to fft_portion_select.py
+- changed fft_portion_select from sample data to any data
+
+- commented-out trimming timeseries if any negative intensity values (led to empty arrays - wasn't necessary?)
+
+4/9 :
+- added 3x3 pixel-box averaging spectra example + 3x3 pixel box visual example to Example_process_6_segments_FFT.py
+
+- added 3x3 pixel-box averaging to fft_overlap in SolSpec.py